requirements.txt
-U django==1.10
pillow
django-registration-redux
django-bootstrap-toolkit